Remove duplicate test case
This was a duplicate test case introduced in PR and commit: https://github.com/activeadmin/inherited_resources/pull/258/files#diff-62c651579bf983dfc4dc2a48c50a3440R470

If you expand the lines on that file and go all the way to around 750 or so and you'll see the same test case name. That version of the test case has additional checks in the end but is the same as this one (which never ran because of the duplicate name). The only difference is the one running has stubs for `persisted?` on table and dish right after the intialization of each.